pyproject.toml
[project]
name = "underwriting-insurance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.3",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.12.0",
    "async-timeout>=5.0.1",
    "attrs>=25.4.0",
    "certifi>=2025.11.12",
    "charset-normalizer>=3.4.4",
    "colorama>=0.4.6",
    "dataclasses-json>=0.6.7",
    "distro>=1.9.0",
    "exceptiongroup>=1.3.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "fpdf>=1.7.2",
    "frozenlist>=1.8.0",
    "greenlet>=3.3.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna>=3.11",
    "jiter>=0.12.0",
    "joblib>=1.5.3",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "langsmith>=0.6.0",
    "marshmallow>=3.26.2",
    "multidict>=6.7.0",
    "mypy-extensions>=1.1.0",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "orjson>=3.11.5",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "propcache>=0.4.1",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rasa>=0.1.1",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "scikit-learn>=1.8.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.15.0",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.2",
    "urllib3>=2.3.0",
    "uvicorn>=0.40.0",
    "yarl>=1.22.0",
    "streamlit>=1.32.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pyzmq>=27.1.0",
    "jupyter>=1.0.0",
    "pytest>=9.0.2",
]
[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["underwriting_insurance"]
